structure the project
