+++ Gemfile
@@ -1,0 +1,7 @@
+source "https://rubygems.org"
+
+gem "sinatra"
+gem "sequel"
+gem "erubis"
+gem "sqlite3"
+gem "mechanize"
