--- batch.sh
@@ -1,12 +1,0 @@
-#!
-
-cd /home/admin/bestgems
-ruby crawl.rb
-ruby daily.rb
-ruby total_rank.rb
-ruby daily_rank.rb
-ruby featured.rb
-ruby stat_num_of_gems.rb
-ruby stat_total_downloads.rb
-ruby stat_daily_downloads.rb
-ruby update_date.rb
