+++ bin/batch.sh
@@ -1,0 +1,12 @@
+#!
+
+cd /home/admin/bestgems
+ruby lib/crawl.rb
+ruby lib/daily.rb
+ruby lib/total_rank.rb
+ruby lib/daily_rank.rb
+ruby lib/featured.rb
+ruby lib/stat_num_of_gems.rb
+ruby lib/stat_total_downloads.rb
+ruby lib/stat_daily_downloads.rb
+ruby lib/update_date.rb
