+++ bin/migrate
@@ -1,0 +1,2 @@
+#!
+sequel -m migrations sqlite://db/master.sqlite3
