+++ lib/create_report.rb
@@ -1,5 +1,5 @@
 require 'sequel'
-DB = Sequel.sqlite('DB', :timeout => 60000)
+DB = Sequel.sqlite('db/master.sqlite3', :timeout => 60000)
 reports = DB[:reports]
 
 p reports.insert({:name => ARGV[0], :summary => ARGV[1], :url => ARGV[2]})
