+++ lib/date2date.rb
@@ -5,5 +5,5 @@ start_date = Date.parse(ARGV[0])
 end_date = Date.parse(ARGV[1])
 
 (start_date..end_date).each{|date|
-	puts date.strftime("%Y-%m-%d")
+  puts date.strftime("%Y-%m-%d")
 }
