+++ lib/db.rb
@@ -1,6 +1,6 @@
 require 'sequel'
 
-DB = Sequel.sqlite('DB', :timeout => 60000)
+DB = Sequel.sqlite('db/master.sqlite3', :timeout => 60000)
 total = DB[:total]
 daily = DB[:daily]
 master = DB[:master]
