+++ lib/stat.rb
